requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.10
python-dotenv==1.0.1

google-genai
discord.py==2.3.2
pyyaml==6.0.1
lxml==5.1.0
playwright==1.41.0
tweepy==4.14.0
atproto>=0.0.65
